Setting default config in the root gradle
