+++ android-plugin/src/functionalTest/java/com/elastic/apm/android/plugin/testutils/BaseFunctionalTest.java
@@ -59,7 +59,7 @@ public abstract class BaseFunctionalTest {
     }
 
     protected ElasticBlockBuilder getDefaultElasticBlockBuilder() {
-        return buildFileBuilder.getAndroidBlockBuilder().getDefaultConfig().getElasticBlockBuilder();
+        return buildFileBuilder.getElasticBlockBuilder();
     }
 
     protected void setUpProject() {
